+++ src/core/relaxation_scheme.jl
@@ -190,3 +190,50 @@ function relaxation_product_on_off(m, x, y, z, ind)
     @constraint(m, z <= x_ub*y + y_lb*x - ind*x_ub*y_lb)
 end
 
+"""
+convex hull relaxation of trilinear term
+
+```
+w₁ = getlowerbound(x)*getlowerbound(y)*getlowerbound(z)
+w₂ = getlowerbound(x)*getlowerbound(y)*getupperbound(z)
+w₃ = getlowerbound(x)*getupperbound(y)*getlowerbound(z)
+w₄ = getlowerbound(x)*getupperbound(y)*getupperbound(z)
+w₅ = getupperbound(x)*getlowerbound(y)*getlowerbound(z)
+w₆ = getupperbound(x)*getlowerbound(y)*getupperbound(z)
+w₇ = getupperbound(x)*getupperbound(y)*getlowerbound(z)
+w₈ = getupperbound(x)*getupperbound(y)*getupperbound(z)
+w = λ₁*w₁ + λ₂*w₂ + λ₃*w₃ + λ₄*w₄ + λ₅*w₅ + λ₆*w₆ + λ₇*w₇ + λ₈*w₈
+x = (λ₁ + λ₂ + λ₃ + λ₄)*getlowerbound(x) + (λ₅ + λ₆ + λ₇ + λ₈)*getupperbound(x)
+y = (λ₁ + λ₂ + λ₅ + λ₆)*getlowerbound(x) + (λ₃ + λ₄ + λ₇ + λ₈)*getupperbound(x)
+z = (λ₁ + λ₃ + λ₅ + λ₇)*getlowerbound(x) + (λ₂ + λ₄ + λ₆ + λ₈)*getupperbound(x)
+λ₁ + λ₂ + λ₃ + λ₄ + λ₅ + λ₆ + λ₇ + λ₈ = 1
+```
+"""
+function relaxation_trilinear(m, x, y, z, w, lambda)
+    x_ub = getupperbound(x)
+    x_lb = getlowerbound(x)
+    y_ub = getupperbound(y)
+    y_lb = getlowerbound(y)
+    z_ub = getupperbound(z)
+    z_lb = getlowerbound(z)
+
+    @assert length(lambda) == 8
+
+    w_val = [x_lb * y_lb * z_lb 
+             x_lb * y_lb * z_ub  
+             x_lb * y_ub * z_lb
+             x_lb * y_ub * z_ub 
+             x_ub * y_lb * z_lb
+             x_ub * y_lb * z_ub
+             x_ub * y_ub * z_lb
+             x_ub * y_ub * z_ub]
+
+    @constraint(m, w == sum(w_val[i]*lambda[i] for i in 1:8))
+    @constraint(m, x == (lambda[1] + lambda[2] + lambda[3] + lambda[4])*x_lb + 
+                        (lambda[5] + lambda[6] + lambda[7] + lambda[8])*x_ub)
+    @constraint(m, y == (lambda[1] + lambda[2] + lambda[5] + lambda[6])*y_lb + 
+                        (lambda[3] + lambda[4] + lambda[7] + lambda[8])*y_ub)
+    @constraint(m, z == (lambda[1] + lambda[3] + lambda[5] + lambda[7])*z_lb + 
+                        (lambda[2] + lambda[4] + lambda[6] + lambda[8])*z_ub)
+    @constraint(m, sum(lambda) == 1)
+end